+++ tendermint-lite/src/lib.rs
@@ -1,11 +1,1 @@
-
-
-pub mod lite;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+pub mod requester;
